YUK-8486 CCU Simulator - Change the error chances to be doubles instead of ints Updated the delay behavior, commsbehavior, and CCU Simulator files to use the cparm values of doubles instead of integers.

+++ yukon-server/Simulator/include/CommsBehavior.h
@@ -11,7 +11,7 @@ public:
     typedef bytes target_type;
 
     virtual void apply(target_type &message)=0;
-    virtual void setChance(int chance)=0;
+    virtual void setChance(double chance)=0;
 
     enum BehaviorValues
     {
