+++ yukon-server/Simulator/include/DelayBehavior.h
@@ -11,11 +11,11 @@ class DelayBehavior : public CommsBehavior
 public:
     DelayBehavior();
     virtual void apply(bytes &message);
-    void setChance(int chance);
+    void setChance(double chance);
 
 private:
     bytes _delayed;
-    int _chance;
+    double _chance;
 };
 
 }
